Update to 1.7, add ES6 app, include config.xml
Use Tabris.js 1.7 syntax (widget constructors).
Update version to match with the tabris module.

Instead of an empty and a "native widget totally rock" app,
offer the creation of a Basic app and an ES6 app. Include the
"native widget totally rock" example in both apps, this seems
to be a good start.

Optionally create a config.xml file, it's required by the
build service.

Remove the question for the entrypoint name to avoid asking too
many questions. There doesn't seem to be much need to customize
this file name - "app.js" is as good as any other name.

Add LICENSE file.

Change-Id: I0517044ddd618dc64754846ecf93013d492b495d