Add ESLint config
Use ESLint [1] to ensure code quality.

Prefix package.json in template with underscore to prevent ESLint from
trying to read it (and fail).

Move templates for basic and ES6 example to separate folders basic/ and
es6/, respectively. This clears up the structure and it allows for
different eslint configs.

Add npm script "lint" to run the linter with:

    npm run lint

[1] http://eslint.org/