Upgrade to Yeoman v1.1.0
Generator.prompt() returns a promise now, which resolves with "answers".
prompting() is handled asynchronously when returning a promise.

Rename "build" to "compile" because it was not clear what the difference
between "proj_build" (Cordova build) and "build" was (ES6/TS
compilation). Also don't expose it as a template property, because it is
not used in any template.

Change-Id: Ie87d778cc2b42198633d516eca55674828729094