Add linter config for JS and TS projects
Add basic ESLint [1] config file for JavaScripts projects and a
TSLint [2] config file for TypeScript projects.

Npm-install the respective linter after generation.

Place the config files in the template directories with their
original names to ensure the source code templates are checked.

Add a dependency to tslint and typescript to lint the ts template.

[1] http://eslint.org/
[2] http://palantir.github.io/tslint/

Fix #30

Change-Id: Ia196d49b917715e71deacad06da34d18f9e1ca1b