Fail with error message on old Node.js versions
When trying to run ES6 code on old Node.js versions, the resulting
error message usually points at some lexical error in a source file.

Provide a clear error message to point users with outdated Node.js
versions in the right direction.

Replace index.js with a ES5-compatible delegator module that performs
the version check.

A test would be hard to write as the test suite itself is based on ES6.