Reduce number of questions asked
The generator should quickly create a new project. To speed up this
process, remove the questions for Author, Email, and App description.
These meta data are not required and they can still be changed in the
config.xml before publishing the app.

Use the git username and email address to fill in the defaults.

Change-Id: I1e055e481b9ae48fa271f0d6e08217cdd4bf7b44