Provide separate templates for Tabris 2.x and 3.x apps
While it would be possible to share some files between the two, this
approach reduces the risk of breaking the 2.x template (which will
likely not change much anymore) while revising the 3.x template.

The new templates are slightly modernized versions to the previous ones.
In case of the TypeScript 3.x template wrap the code in a class to
encourage coding like in a high-level language. A simpler non-TypeScript
JSX template (akin to the 2.x one) could be added later.

Tweak the wording of the generator questions for clarity.

Change-Id: I9621a1f307e3d1d3fcbfe05d54e62af0c20d1528