Install "latest" patch version of tabris.js for 3.x projects
The release process for 3.x now publishes the tabris module before the
platforms and developer app, but with the dist-tag "rc" instead of
"latest". This module version is not to be installed until it is tagged
"latest", so use that tag for module installation. To prevent npm from
deciding by itself to upgrade to new minor version it needs to be
specified that it should stick to the same patch version using '~'.

Also stick to the TypeScript compiler minor version we use in the tabris
build.

Remove pre-release related code.

Set version to 3.0

Change-Id: Ib6f1ef2aa0682dca4f825bbeaf384ff98d04d64e