Migrate templates from tslint to eslint
Rules are relaxed variants of the eslintrc used for the generator
itself. Same plug-ins are used as well.

A tsconfig.test.json is now included in projects that have mocha tests.
It is required by the eslint TypeScript plugin.

Tests have been reworked to actually lint the generated projects. To do
so the npm modules of these projects have to be installed, but they are
re-used for each initial code variant to save time.

Change-Id: I692913267269a1ab3fd4aabd219c9289a994bc75